pyproject.toml
[tool.poetry]
name = "project management"
version = "0.1"
description = "Software to manage between geographical spread teams"
authors = ["Ahmad"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.1"
fastapi = "0.112.1"
sqlalchemy = "^2.0.21"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
aiosqlite = "^0.19.0"
jinja2 = "^3.1.2"
httpx = "^0.25.0"
nh3 = "^0.2.14"
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"
minio = "^7.2.7"
fastapi-csrf-protect = "^0.3.4"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
prompt-toolkit = "^3.0.47"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.291"
black = "^23.9.1"
isort = "^5.12.0"
loguru = "^0.7.2"
ipykernel = "^6.25.2"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.11.1
target-version = "py311"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10